requirements.txt
smbus2
RPi.GPIO
click
RPi.bme280
urllib3
python-dateutil
